Fixes for 32-bit musl libc
musl needs a specific include for "struct timeval" and uses
"long long int" for time_t on 32bit. Casting time_t to "long" should
be fine for the time being, we still have almost 17 years.

+++ log.c
@@ -64,7 +64,7 @@ __msg(int lvl, const char *func, const char *format, ...)
 
 		clock_gettime(LOG_CLOCK, &now);
 		snprintf(timebuf, sizeof(timebuf), "%6ld.%06ld",
-			 now.tv_sec, now.tv_nsec / 1000);
+			 (long)now.tv_sec, now.tv_nsec / 1000);
 	} else
 		*timebuf = '\0';
